Add files via upload
Genesis. Initial test works. Running on CPU. Reducing mutation rate and character equality fitness from 0.1 to 0.01 yields better results. Struggled with the roulette wheel algorithm, and was causing some segfaults because of it's inaccuracy. Performance can be improved. It's currently n^2 and could be n*log(n). A sort and a binary search are in order; the problem is keeping the fitness array mapped to the initial population array. I'm also unsure if the upper bound of the fitness array could overflow, causing no candidates to be selected, and an empty offspring in it's place - which was the main cause for the segfaults.
Paralellizing with CUDA is a potetntial option. CMakeLists.txt includes support for CUDA and OpenGL, to also visualize the results better.